Content Style: Width
The content style struct now stores a width. This allows row headers to
have their column breaks specified as part of the cells in the header
row.

This also opens the door for cells to specify their desired widths to
support more complex future rendering cases such as row headers.