ColumnBreak -> CellWidth: Part of content style
Removes the ColumnBreak struct. This is replaced by the CellWidth
struct, which is a part of the content style module.

The structure had been moving in this direction for a while. Header rows
have cells in them, which declare the width of columns in the table. It
makes sense for the content of the header row to be defined within the
cell style directives. It followed that it made more sense to have the
cell width formally be a part of the content style, thereby removing the
need for a column break struct.